Align with PR #201: fix KEEPALIVE_V2_ACK to 28 bytes, add KeepAliveV2AckFrame, correct field names

+++ src/protocol/src/protocol/solo.cpp
@@ -1021,15 +1021,15 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
             packet.m_header, get_llp_header_name(packet.m_header), packet.m_length);
         return;
     }
-
+    
     /* Guard: Un-mirrored stateless opcodes must never arrive on legacy lane */
-    if (m_protocol_lane == ProtocolLane::LEGACY &&
-        ::LLP::IsUnmirroredDataOpcode(static_cast<uint16_t>(packet.m_header)))
+    if(m_protocol_lane == ProtocolLane::LEGACY &&
+       ::LLP::IsUnmirroredDataOpcode(static_cast<uint16_t>(packet.m_header)))
     {
         m_logger->error("[Colin] REJECTED un-mirrored stateless opcode 0x{:04x} ({}) on legacy lane"
                         " — stateless port required",
             packet.m_header, ::LLP::GetUnmirroredOpcodeName(static_cast<uint16_t>(packet.m_header)));
-        if (connection) connection->close();
+        if(connection) connection->close();
         return;
     }
     
@@ -2857,7 +2857,7 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
     else if (matches_opcode(0xE0))
     {
         /* PING_DIAG is stateless-only — reject on legacy lane */
-        if (m_protocol_lane != ProtocolLane::STATELESS)
+        if(m_protocol_lane != ProtocolLane::STATELESS)
         {
             m_logger->warn("[Colin PING] PING_DIAG rejected on non-stateless lane — stateless port required");
             return;
@@ -2866,7 +2866,7 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
         /* Exact payload size enforcement for fixed-size PING_DIAG opcode */
         std::vector<uint8_t> payload = packet.m_data ? *packet.m_data : std::vector<uint8_t>{};
         uint32_t nExpected = ::LLP::GetExpectedPayloadSize(::LLP::ColinDiagOpcodes::PING_DIAG);
-        if (payload.size() != nExpected)
+        if(payload.size() != nExpected)
         {
             m_logger->warn("[Colin PING] Payload size mismatch for PING_DIAG:"
                            " expected {} bytes, got {} — discarding",
@@ -2876,7 +2876,7 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
 
         /* Parse 64-byte PingFrame, build telemetry-enriched PongFrame, reply immediately */
         auto pong_bytes = m_colin_ping_handler.HandlePing(payload, true /* stateless */);
-        if (!pong_bytes.empty() && connection)
+        if(!pong_bytes.empty() && connection)
         {
             /* PONG opcode: 0xD0E1 stateless (mirror-mapped by PacketBuilder) */
             connection->transmit(PacketBuilder::build(m_protocol_lane, 0xE1, pong_bytes));
@@ -2887,12 +2887,13 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
     // ═══════════════════════════════════════════════════════════════════════
     // KEEPALIVE_V2_ACK (0xD101) — stateless-only, 28-byte chain-state payload
     // ═══════════════════════════════════════════════════════════════════════
-    else if (packet.m_is_uint16_opcode &&
-             packet.m_header == static_cast<uint16_t>(::LLP::KeepAliveV2Opcodes::KEEPALIVE_V2_ACK))
+    else if(packet.m_is_uint16_opcode &&
+            packet.m_header == ::LLP::KeepAliveV2Opcodes::KEEPALIVE_V2_ACK)
     {
+        /* Exact payload size enforcement for KEEPALIVE_V2_ACK (28 bytes) */
         std::vector<uint8_t> payload = packet.m_data ? *packet.m_data : std::vector<uint8_t>{};
         uint32_t nExpected = ::LLP::GetExpectedPayloadSize(::LLP::KeepAliveV2Opcodes::KEEPALIVE_V2_ACK);
-        if (payload.size() != nExpected)
+        if(payload.size() != nExpected)
         {
             m_logger->warn("[KEEPALIVE_V2] ACK payload size mismatch:"
                            " expected {} bytes, got {} — discarding",
@@ -2900,7 +2901,7 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
             return;
         }
         ::LLP::KeepAliveV2AckFrame ack;
-        if (ack.Parse(payload))
+        if(ack.Parse(payload))
         {
             m_logger->debug("[KEEPALIVE_V2] ACK received: seq={}"
                             " unified_height={} prime_height={} hash_height={}"
@@ -2909,8 +2910,8 @@ void Solo::process_messages(Packet packet, std::shared_ptr<network::Connection> 
                 ack.unified_height, ack.prime_height, ack.hash_height,
                 ack.hashPrevBlock_lo32, ack.hash_tip_lo32, ack.fork_score);
 
-            /* Fork detection: compare node's chain tip against the miner's prevHash canary */
-            if (ack.IsForkDetected(ack.hashPrevBlock_lo32))
+            /* Fork detection: compare node's chain tip against miner's prevHash canary */
+            if(ack.IsForkDetected(ack.hashPrevBlock_lo32))
             {
                 m_logger->warn("[KEEPALIVE_V2] Fork detected!"
                                " miner_prevHash_lo32=0x{:08x} node_tip_lo32=0x{:08x} fork_score={}",
